commit 6
The layout of the header on the admin side has been modified this time. Change the table on the admin homepage to a mat table with the material theme. Add titles to the add and edit interfaces in admin. Fixed a bug in the admin homepage where the home content is not displayed when using API/admin. When no ID is selected on the edit interface, a prompt message pops up.